Repository che legge da tabella

+++ src/main/java/it/intesys/codylab/model/Cerchio.java
@@ -2,20 +2,20 @@ package it.intesys.codylab.model;
 
 public class Cerchio implements FormaGeometrica{
 
-    private final float raggio;
+    private final double raggio;
 
-    public Cerchio(float raggio) {
+    public Cerchio(double raggio) {
         this.raggio = raggio;
     }
 
     @Override
-    public float perimetro() {
-        return Double.valueOf(Math.PI * 2 * raggio).floatValue();
+    public double perimetro() {
+        return Double.valueOf(Math.PI * 2 * raggio).doubleValue();
     }
 
     @Override
-    public float area() {
-        return Double.valueOf(Math.PI * raggio * raggio).floatValue();
+    public double area() {
+        return Double.valueOf(Math.PI * raggio * raggio).doubleValue();
     }
 
     @Override
