+++ src/main/java/it/intesys/codylab/model/FormaGeometrica.java
@@ -2,6 +2,6 @@ package it.intesys.codylab.model;
 
 public interface FormaGeometrica {
 
-    float perimetro();
-    float area();
+    double perimetro();
+    double area();
 }
