+++ src/main/java/it/intesys/codylab/model/Quadrato.java
@@ -2,15 +2,15 @@ package it.intesys.codylab.model;
 
 public class Quadrato extends Quadrilatero {
 
-    private final float lato;
+    private final double lato;
 
-    public Quadrato(float lato) {
+    public Quadrato(double lato) {
         super(lato, lato, lato, lato);
         this.lato = lato;
     }
 
     @Override
-    public float area() {
+    public double area() {
         return lato * lato;
     }
